Change SQLite file extension to .sqlite3
As a user of Sublime Text, Switching to the :sql adaptor with sqlite
generates a `<app_name>_development.db` file. Sublime text by default
hides .db extension, and having the original directories generated by
the :file_system adaptor lying around added a little bit of confusion
as to why I couldn't see a database file, even though I could access it
within the REPL.

Although the adaptor is specified within the codebase, the database
file is probably more likely to be stored in another location outwith
the repository when deployed to a server. Having a more useful extension
isn't a bad thing to have.