Vendor PhantomJS 2.1.1 in our repository
As of today Travis CI supports PhantomJS 1.9.
Previously, the CI script (in `.travis.yml`) tried to fetch it before
each build.

But this is inefficient and the official channel of distribution for
PhantomJS ofter drops the connection and the build fails miseably.

When Travis CI will support PhantomJS 2.1 we can remove it and review
the CI setup again.