Add action extensions for view rendering
This contains a lot of hacks because of the approach of using monkey-patches, but the core mechanic seems sound:

 - Customi build_response method that passes in a specific `view_options` proc for the action
- Which is actually a method object of the private #view_options method (which means this is customizable in specific subclasses if ever needed, yay for flexibility!)
- By default this supplies a view context object with custom request-specific details passed in via a separate #view_context_options method (also customizable in subclasses!)

The end result is that we can just call res.render(view) in an action subclass and the view will be rendered with a _fully featured_ context object (i.e. full of request-specific details) without any extra work on the user’s behalf. It’s just seamless.