Set HANAMI_ENV from CLI -e flag before initing app
This ensures all aspects of the application boot process respect the provided env.

An example use case: an application may have a type argument provided to one of its application settings (i.e. in config/settings.rb) that uses the current `Hanami.env`. If the application is init’ed before the CLI `-e` flag sets the env, then that env will potentially be different when the setting is processed versus when the application finally boots in full.