CLI application env handling and dotenv improvements (#1045)
* Set HANAMI_ENV from CLI -e flag before initing app

This ensures all aspects of the application boot process respect the provided env.

An example use case: an application may have a type argument provided to one of its application settings (i.e. in config/settings.rb) that uses the current `Hanami.env`. If the application is init’ed before the CLI `-e` flag sets the env, then that env will potentially be different when the setting is processed versus when the application finally boots in full.

* Load range of env-specific and local dotenv files

This brings our behavior into line with dotenv-rails

In our initial,simplistic approach, we loaded just a single “.env” file, which is not enough for handling different env-specific values