Fix isolation specs
These broke when inheriting from Hanami::Action resulted in a call to Hanami.application.component_provider, which wants the slices to be registered already.

Explicitly registering the slices earlier and init’ing the app is the fix here, but I’d also like to explore making single file Hanami apps a little more ergonomic in the future.