Use full action ApplicationConfiguration object (#1068)
This changes the approach for exposing application-level action config. Instead of taking responsibility for building up the config object at the framework level, instead defer in full to hanami-controller, and allow it to provide its own appropriate object for application-level configuration, in this case, the new `Hanami::Action::ApplicationConfiguration`.

This gets us closer to the ideal framework extension approach, too – less hard-coding at the framework level, more delegation to the integrating extensions. Eventually we'll provide some extension point for other gems to provide their own application-level config, but until then, this gets us close to that model in spirit.