Pass base config to actions; finalize all configs
This allows the actions configuration to adapt to various top-level configs, e.g. to enable CSRF protections automatically in the case of `config.sessions.enabled?` being true.

We handle this by passing the base application config to the nested actions config, and then call `finalize!` on the nested actions config when the base config is finalized.

This means the nested actions config can adapt its own configuration at the last moment, which ensures we support the user specifying their configs in whatever order they like prior to this.