Use dry-configurable for hanami settings
This is a baby step towards the to-be-defined level of integration we
want between hanami settings and dry-configurable.

This commit introduces the minimal changes to keep the very same
features we're supporting on the hanami settings side while still
delegating to dry-configurable what it's already capable of handling:

- dry-configurable is responsible for the setter/getter methods.
- Hanami settings resolve the value through a type constructor, when
  present, before delegating it to dry-configurable.
- Hanami settings wrap the definition of all the settings to detect all
  errors at once instead of failing on a one-by-one basis.

Further integration can include delegating the type constructor to a
processor keyword argument (we could already delegate it to the block,
but probably we want to do the former) and provide an error handler to
the whole definition.