Use a class for the application settings
We switch from a block-based definition for settings to a regular Ruby
class so that users can leverage all the flexibility that it brings,
like defining custom methods or using whatever Types module they want.
We also get rid of custom configuration code and depend on dry-configurable.

By default, the settings class is expected to be defined in
config/settings.rb file, and its name should be MyApp::Settings,
where MyApp represents the application namespace.

The user settings class inherits from a Hanami::Application::Settings
class, a dry-configurable-enabled class that delegates any unknown
method to its Config instance so that users don't need to define
reader: true every single time.

We're also breaking apart a settings store abstraction, which was previously
coupled with the load process. By default, we use a store that leverages Dotenv to
load settings from .env files, or directly from ENV when not available.
The loading of settings happens when the instance is initialized.

We leverage the parent class Hanami::Application::Settings to
use it as a null object when users don't define their own class.

Co-authored-by: Tim Riley <tim@riley.id.au>