Use a class for the application routes (#1116)
We move the definition of routes as a block provided to
`Hanami::Application.routes` to its own class
`Hanami::Application::Routes`.

Users need to create a subclass and define the routes through a block
applied to the `.define` class method. As before, the hanami application
captures the context of the block and forwards it to
`Hanami::Application::Router` for its evaluation.

E.g.:

```ruby

require "hanami/application/routes"

module MyApp
  class Routes < Hanami::Application::Routes
    define do
      slice :main, at: "/" do
        root to: "home.show"
      end
    end
  end
end
```

For consistency with settings, we move the routes path configuration to
the main `Hanami::Configuration` class.