Make a routes helper class available in the container (#1119)
* Make a routes helper class available in the container

We add a new `Hanami::Application::RoutesHelper` class which delegates
`path` & `url` methods into the router's URL helpers. The intention is
for it to be made available to whichever view system in use (i.e.,
dry-view).

We have slightly rearranged the boot process. Instead of explicitly
creating the router when the application instance is initialized, we add
it to the `boot` phase of the application (which is effectively the
same). This way, we can register the helpers in the container just
afterward, still on the boot phase, when the container is finalized.
There's no need to load the routes beforehand as the router will already
do that.

* Update lib/hanami/application/routes_helper.rb

Co-authored-by: Luca Guidi <me@lucaguidi.com>