Allow slices to configure their container exports
Right now, we've supported this in the simplest possible way by allowing the container to be directly configured within the slice config block.

Longer term, if we expect this to be a common task (which I think it probably will), it may make sense to put something for this at the slice level, rather than expecting our users to know about and delve into the direct container configuration (this will always be accessible, but it should ideally be reserved for advanced use cases only).