Eagerly initialize autoloader and container
It's useful to have these around at all times, even in their pre-prepared state, plus it removes a bunch of checking we have to do to determine whether they should be returned

Also, stop trying to respect a user-defined container. We need the container to be configured per our own requirements.