Introduce Hanami::SliceName (#1159)
This gives us a single object that can do the work of determining a slice/application's name and namespace, which means we can now set this up early and then pass it to any places that need it, like the configuration (since it wants to provide the app name as part of logger configuration).

This is an improvement on the (temporary) solution we had for the logger configuration, which required us to entirely relocate the work of determining the namespace etc. to the configuration object itself, which isn't really its job.