Ensure prepare_container works as expected in App (#1185)
We want to ensure that the `prepare_container` block is run at the _last moment_ before the container is marked as `configured!`. With this change, this is now true for both `App` and `Slice`.

Before this change, the `prepare_container` block for `App` was being run before the app-specific configuration was applied.