Load settings via provider after autoloader setup
This keeps the details of the settings loading inside the provider class, simplifying Slice.

By loading the settings after the autoloader is set up, we also ensure that constants made available by the autoloader can be referenced within the Settings class (like e.g. a Types module). This takes away a potential hitch for users, and makes the settings class much more like any ordinary class in terms of access to the rest of the application code.