Improve LOAD_PATH handling
Add `App.prepare_load_path` as public API for people who have modified the `config.root` and want to ensure the new root’s `lib/` subdir is immediately added to the load path.

Re-run `.prepare_load_path` as part of the `App.prepare` step to ensure the load path is updated in the case of a re-configured app root.