Conditional slice loading (#1189)
Enable conditional slice loading via a new `config.slices` setting.

For example, given this app class:

```ruby
module TestApp
  class App < Hanami::App
    config.slices = %w[admin]
  end
end
```

And the presence of both `slices/admin/` and `slices/main/`, when the app is prepared (or booted), only the admin slice will be loaded:

```ruby
require "hanami/prepare"

Hanami.app.slices.keys # => [:admin]

Admin::Slice # exists, as expected
Main         # raises NameError, since it was never loaded
```

When slices are skipped from loading, they're _completely ignored_ — their slice classes are not loaded, nor are any of their source files. This is valuable because it can keep the app's memory profile as low as possible for any configured set of slices.

Since loading/skipping particular slices is likely to be common adjustment for different deployment modes (e.g. deploying just the admin slice to a different set of containers, or running just a subset of slices for a background worker), we make it easy to specify this config via an `HANAMI_SLICES` env var. When using this, specify the slices as comma-separated:

```
HANAMI_SLICES="main,shop"
```

Lastly, the `config.slices` setting also supports arbitrarily nested slices. For example, if have an "admin" slice with its own child "shop" and "editorial" slices, you can specify `config.slices.load_slices = ["admin.shop"]` to load only its nested shop slice.

One caveat about this feature is that it's not "smart" — if one slice e.g. imports another (or expects another slice to be available for any other reason), and if that other slice is skipped, you'll receive errors. It's entirely the responsibility of the user to make sure that the appropriate set of slices is loaded when configuring `load_slices`/`skip_slices`.