Raise when trying to configure a missing component
Up to this point, we were returning a `NullConfiguration` instance when
trying to configure an unavailable component. That was too implicit. We
want to tell users what is missing.

That also lets us remove the check for rack when creating a web
application and detect the wrong assumption earlier.

We take the occasion to add missing tests for the router configuration.