Make router configuration optional and have the config system raise when a component is missing (#1204)
* Configure the router only when hanami-router is present

That's the same we're doing with views and actions

* Raise when trying to configure a missing component

Up to this point, we were returning a `NullConfiguration` instance when
trying to configure an unavailable component. That was too implicit. We
want to tell users what is missing.

That also lets us remove the check for rack when creating a web
application and detect the wrong assumption earlier.

We take the occasion to add missing tests for the router configuration.

* Check if gem is bundled when loading components configuration

We leverage `Hanami#bundled?` instead of rescuing from the corresponding
`LoadError`.

By passing the gem name and the config method we're also able to render
a more accurate error message when it's not present.