Make settings available in app/slice class bodies (#1213)
This change restores class-level access to `settings` within app and slice classes. This is important because it allows user-defined settings to be used to configure the app itself. For example:

```ruby
module MyApp
  class App < Hanami::App
    config.sessions = :cookie, {
      key: "my_app.session",
      secret: settings.session_secret, # <=== here
      expire_after: 60*60*24*365
    }
  end
end
```

To support this, we:

- No longer use a container provider to load the settings, since providers can only be used after the slice `prepare` step, which is too late for immediate access inside the class body.
- Remove the expectation that autoloadable constants are available inside the settings classes, since autoloading is also only available after the slice `prepare` step.
- Assist with easy use of type objects for coercing/validating settings by auto-generating a nested `Types` constant within `Hanami::Settings` subclasses. This is intended to replace the use of the `MyApp::Types` module defined in `lib/my_app/types.rb` or `app/types.rb`.

A note about the types module:

Since types for settings tend to be simple, this kind of inline types definition should be sufficient for the majority of cases, and is also the easiest for us to describe in documentation and other supporting material.

Users requiring more advanced usage of types (such as sharing of types between the settings classes and elsewhere) can do so by adding their own manual `require` statements to the class as needed. By doing this and eschewing the provided nested types module, this is a clear signal to the user that they're going off the "golden path", and as such, their expectations around support via documentation etc. should be adjusted accordingly.