Raise errors for missing settings (#1222)
In https://github.com/hanami/hanami/pull/1216 (specifically, https://github.com/hanami/hanami/pull/1216/commits/d4c036b32a4471a13d70a15bc7f8cef9b5c4509f), we made a small tweak to our settings loading to adapt to changed handling of `Undefined` values in dry-configurable 0.16.0.

The change in dry-configurable is that `Undefined` values sent to config no longer implicitly fall back to using their default value; `Undefined` instead is accepted as the value itself (though not run through the constructor).

To work around this, we stopped setting values on our Settings config object if no such key existed in the store (i.e. if no ENV var existed at all).

However, this turned out to have the undesirable side effect of missing ENV vars no longer triggering exceptions from their constructors. That is, if a setting is required, but not defined in ENV at all, it would no longer halt the boot of the Hanami app and provide early feedback to the user.

This is critical behavior for Hanami, so here we restore it through attempting to _read_ the config value for any keys missing from ENV. This ensures the consrtuctor is run (as it is done for all config values read for the first time) and errors are appropriately raised.

Add a new unit test for this specific scenario at the same time.

Also, once the config has finished loading, finalize it so that its values cannot subsequently be changed.