Clarify docs for environment
This makes it clearer that the argument yielded to the block should be used to set the config. Doing so is safer than relying on the implicit `self`, since it responds to `#config` because of its internal `#config` reader, but this is not the same object as the config itself.