Simplify and clarify usage of Config#environment (#1238)
Since 4a085a5, the `env` cannot be changed for an `Hanami::Config`. This means we do not need to store the environment-specific config blocks, and can instead evaluate them immediately. This is simpler for the user to understand and makes for a much simpler internal implementation too.

Also update the docs to make it clearer that the argument yielded to the `#environment` block should be the object used to set the config. Using this yielded argument is safer than relying on the implicit `self`, since because `Config#config` exists, a user may call `config` inside the block (with implicit self) and receive the internal `Dry::Configurable::Config` when they actually expect `Hanami::Config`.