Fix action/view integration when hanami-view is not bundled (#1313)
This should have been checking explicitly for whether hanami-view was bundled, rather than checking for the presence of the Extensions::Views const, which is automatically made available via Zeitwerk, and led to a LoadError when hanami-view was not available.

The only need we had for our action extension to reach into the view extension was for the shared code for loading the view context. Instead, we now explicitly check to see if hanami-view is bundled before using that code.

In addition, we add a fallback to look for a same-named context class in the conventional location, which would be useful if a 3rd party view layer was to be used.