Return plain HTTP errors when not rendering errors
When both render_errors *and* render_detailed_errors are disabled, return the default responses from hanami-router for not found errors (HTTP 404) and not allowed errors (HTTP 405).

This fixes an bug where not found errors were being rendered as HTTP 500 responses, showing the stack trace for a Hanami::Router::NotFoundError exception.