Fix error logging on handled exceptions (#1337)
Only log unhandled exceptions as errors.

To do this, adjust the overloaded `#_handle_exception` in our framework-level `Hanami::Action` extension. `#_handle_exception` method is called for _any_ error raised inside an action, regardless of whether it is configured to be handled by the user. So instead of instrumenting (and therefore logging) every exception, only do so if an exception is _raised_ from within this method. This ensures that only true, unhandled exceptions get logged as errors.