Refactor `SliceRegistrar#load_slice` to avoid an exception (#1345)
Currently we are trying to require a file that we aren't sure exists and
rescuing the `LoadError` that we get if it doesn't. 

We can completely avoid an exception per slice for the knowable case
where this file doesn't exist by first checking using `Pathname#sub_ext`
to add the `RB_EXT` extension to the path and then making use of
`Pathname#file?` on the resulting path.