Render a welcome page when no routes are defined (#1353)
The page includes the current Hanami and Ruby versions. Design and content for this page will be added in a later change.

The page is rendered by a new `Hanami::Web::Welcome` middleware that is added early in the stack when (a) in development mode, and (b) no routes are defined.

To support this introspection into the routes (which are passed to the router as an opaque block), wrap the routes in a `RoutesProc` class that adds an `#empty?` method, returning true when no routes are defined.