Use action view context in spirit of hanami-view
In changes to hanami-view soon to be released as 2.1.0, we removed the `Hanami::View::Context#with` method, instead expecting that a view context need only be initialized once, and then kept for the entire rendering.

However, at the time of that change, we did not make commensurate changes here in the app-level view context extension, instead just keeping a local `#with` class, and continuing to (needlessly) initialize a view contexet to pass to the action, and then re-initialize it with the request before rendering.

With this change, make the view app-level view context work in a way that matches the standard Hanami::View::Context: it should only be initialized once. To achieve this, inject a view context *class* into actions (instead of an instance) and then call `.new` on that context class when preparing the options for view rendering.

This simplifies the view context extension internals and allows us to remove a few tests, too.