Use action view context in spirit of hanami-view (#1359)
In earlier changes to hanami-view, we removed the `Hanami::View::Context#with` method, instead expecting that a view context need only be initialized once, and then kept for the entire rendering.

However, at the time of that change, we did not make commensurate changes here in the app-level view context extension, instead just keeping a local `#with` method, and continuing to (needlessly) initialize a view contexet to pass to the action, only to _re-initialize it_ inside the action (via `#with`) with the request before rendering the view.

With this change, make the view app-level view context work in a way that matches the standard `Hanami::View::Context`: it should only be initialized once. To achieve this, inject a view context *class* into actions (instead of an instance) and then call `.new` on that context class when preparing the options for view rendering.

This simplifies the view context extension internals and allows us to remove a few tests, too. It will also make view contexts easier for users to modify, since they no longer have to think about the case of `#with` being called somewhere deep inside the framework, out of their sight. Instead, their context classes will be initialized just once, like any ordinary Ruby object. And the only expectation we have is that their `#initialize` accepts `**kwargs`.