Allow slice definitions to also exist in `slices/[slice_name]/config/slice.rb` (#1347)
Currently the slice has no way to provide its own `Slice` class, but when working with slices it is very helpful to be able to define one so that configuration can be set without expecting the app to do it for the slice.

This change allows a slice to have its class defined inside the slice at a predefined location of `slices/[slice_name]/config/slice.rb`. This works for slices nested under the main app as well as slices nested inside other slices.

If a slice class is defined (a) in the app, or (b) in its parent slice, that will be chosen in preference to a slice class within the slice itself. This means that everything can be re-configured by a slice's parent slice or the app.