Render detailed errors in development mode only
If these get rendered in test mode, then the text in the error messages themselves may actually confuse assertions in the tests, especially for feature tests that assert on page content.

For example, given an expectation like this:

expect(page).to have_content(“Hello world”)

If the page has some error and the BetterErrors screen shows, then it will include the exact copy of this line above as part of its backtrace, which will lead to the test passing due to the test line itself having the text it was looking for.

This is undesirable, and we can avoid any potentially confusing errors like this by simply disabling the detailed errors middleware altogether in test mode.