Pass through kwargs to middleware
Most middleware expect the old-school approach of hash options, but
some are now using keyword arguments instead (such as Rack::Timeout).
So let's presume that any keyword arguments that aren't known by
Hanami should be passed through.

There is the edge-case of both the middleware and Hanami expecting
arguments with the same name - this does not address that.