Do not auto-register from db, relations, structs
db/ and relations/ are special directories for ROM, and either contain files that should not be registered (like db/changesets/, db/commands/, db/mappers.), and relations/ is already managed by the db provider.

strcuts/ (like entities/, which was already handled in this way) contains classes that should be instantiated directly with specific data, instead of being accessed pre-built off the container.