Remove #any_adapter support
This is complicating how I think about applying gateway config from parents, and in general I’ve come to think that this is a step too far for us in our first release of the DB layer. If we need more sophisticated adapter configuration in the future, we can consider reintroducing something like this.