Use gemspec to load json gem with no warnings (#1469)
Version 2.7.2 of the json gem fixes its usage of ostruct so that this warning is no longer emitted on Ruby 3.3.5 and later:

```
/path/to/installs/ruby/3.3.5/lib/ruby/3.3.0/json/generic_object.rb:2: warning: ostruct was loaded from the standard library, but will no longer be part of the default gems starting from Ruby 3.5.0.
```

However, 3.3 versions of Ruby bundle the json gem at version 2.7.1, which does not include this fix. To ensure we use the right version and stop this warning from showing to our users, require the appropriate version of json gem via hanami's own gemspec.

This adds another gemspec dependency, but it's very highly likely that users will already have json bundled anyway.