Only load rom extension if rom is bundled
Move the require for dry-operation’s “dry/operation/extensions/rom” extension to after our checks for the presence of the “hanami-db” gem.

Without this, a hanami app generated with `--skip-db` will fail to boot, because the dry-operation extension raises an error at require-time if the rom-sql gem isn’t available.