Update error message assertion for latest Rack
With the security fix released in Rack 2.2.13, the Rack::Static middleware exits earlier when given
bad paths (e.g. paths trying to escape the root directory via "..", which this test exercises). This
means the Hanami app ends up handling the 404 error, rather than `Rack::Files`, which is what was
happening for previous versions of Rack.