Avoid early loading of routes file (#1539)
When determining whether to load the `routes` provider, check for whether a routes file (or class) exists, but do not attempt to load the file.

Before this change, the `config/routes.rb` file was being loaded as a side effect of checking whether to register the `routes` provider.

Because the providers are registered quite early in the slice boot process, even before the autoloader is set up, this meant that if there are any class references in the routes file, they will end up as errors, because they cannot yet be autloaded.