Do not autoload files that may depend on non-installed gems (#1498)
The hanami gem ships with extensions to a range of _other_ first-party Hanami gems, which helps them integrate better when used in the context of a full Hanami app.

We had been loading the source files for those extensions using Zeitwork.

However, if you then do a Zeitwerk eager load (or run `bin/rails zeitwerk:check` if you're a super cool person trying Hanami _inside_ Rails), this would attempt to load those extension files, which could lead to load errors if any of the Hanami gems they depend on are not installed.

With this change, we exclude these extension files from Zeitwerk autoloading, and instead use manual requires. This solves the above problem.